pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-litellm-autoconfig"
version = "0.5.0"
description = "Profile-driven compose and KubeAI deployment compiler for vLLM stacks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "requests>=2.31",
]

[project.scripts]
vllm-stack = "vllm_service.cli:main"

[tool.setuptools]
packages = ["vllm_service", "vllm_service.backends"]
include-package-data = true

[tool.setuptools.package-data]
vllm_service = ["templates/*.yaml", "templates/*.j2"]
